Organize metrics related tests a bit
